requirements.txt
av==10.0.0
blinker==1.6.2
blosc2==2.0.0
celluloid==0.2.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.4
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
Cython==0.29.35
deepdish==0.3.7
easing-functions==1.0.4
filelock==3.12.2
Flask==2.3.2
fonttools==4.40.0
h5py==3.8.0
idna==3.4
imageio==2.31.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kornia==0.6.12
lazy_loader==0.2
lit==16.0.6
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
nibabel==5.1.0
nilearn==0.10.1
nltools==0.4.7
numexpr==2.8.4
numpy==1.25.0
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
py-cpuinfo==9.0.0
py-feat==0.5.1
pycparser==2.21
pynv==0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sympy==1.12
tables==3.8.0
threadpoolctl==3.1.0
tifffile==2023.4.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
Werkzeug==2.3.6
xgboost==1.7.5
